test(BA-7779): move the scenario surface into testutils and split the kit's own tests
The surface never knew a component: an operation is bound from whatever method it is
given, a matcher reads whatever payload it is handed, and a fake is named by its own
class. Only laying rows down needed the manager, so that is the line.

- ai.backend.testutils.typed_scenario holds the surface. `Seed` is generic over what
  writes rows, so the module names no repository.
- bai_kit.manager.seeding holds `creates()`, which overloads on the creator kind and
  hands back a seed bound to the ops repository.
- The surface's own tests move to tests/unit/testutils, where they run without a
  database. What stays beside the kit is what needs the World: the runner's dispatch
  and what the World seeded.
